Gets criado de exercises e categories

+++ Codigo-Fonte/API/ArpeFitness/src/main/java/pt/iade/ArpeFitness/models/repositories/ExerciseCategoryRepository.java
@@ -3,7 +3,5 @@ package pt.iade.ArpeFitness.models.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.iade.ArpeFitness.models.tables.ExerciseCategory;
 
-public interface ExerciseCategoryRepository extends JpaRepository<ExerciseCategory, Integer> {
-
-    
-} 
+public interface ExerciseCategoryRepository extends JpaRepository<ExerciseCategory, Long> {
+}
