+++ Codigo-Fonte/API/ArpeFitness/src/main/java/pt/iade/ArpeFitness/models/repositories/ExerciseRepository.java
@@ -3,6 +3,7 @@ package pt.iade.ArpeFitness.models.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.iade.ArpeFitness.models.tables.Exercise;
 
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {  
-    
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
+
 }
+
